Remove spurious cast

+++ ini/ini_get_valueobj.c
@@ -224,7 +224,7 @@ int ini_get_config_valueobj(const char *section,
             return EOK;
         }
 
-        if ((hash == (unsigned long int)col_get_item_hash(item)) &&
+        if ((hash == col_get_item_hash(item)) &&
             (strncasecmp(col_get_item_property(item, &len), name, name_len) == 0) &&
             (len == name_len)) {
                 TRACE_INFO_STRING("Item is found", name);
